+++ BeatyPlace/src/test/java/by/beaty/place/repository/AppointmentRepositoryRepositoryTest.java
@@ -8,12 +8,38 @@ import by.beaty.place.model.Users;
 import by.beaty.place.model.common.AppointmentStatus;
 import by.beaty.place.repository.config.BaseRepositoryTest;
 import java.math.BigDecimal;
+import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.junit.jupiter.Container;
 
-class AppointmentRepositoryTest extends BaseRepositoryTest {
+class AppointmentRepositoryRepositoryTest extends BaseRepositoryTest {
+
+    @Container
+    static PostgreSQLContainer<?> postgreSQLContainer = new PostgreSQLContainer<>("postgres:15")
+            .withDatabaseName("prop")
+            .withUsername("postgres")
+            .withPassword("postgres")
+            .waitingFor(Wait.forListeningPort().withStartupTimeout(Duration.ofSeconds(60)))
+            .withExposedPorts(5432);
+
+    @DynamicPropertySource
+    static void registerPgProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url",
+                () -> String.format("jdbc:postgresql://localhost:%d/prop", postgreSQLContainer.getFirstMappedPort()));
+        registry.add("spring.datasource.username", () -> "postgres");
+        registry.add("spring.datasource.password", () -> "postgres");
+        registry.add("spring.flyway.url", () -> String.format("jdbc:postgresql://localhost:%d/prop",
+                postgreSQLContainer.getFirstMappedPort()));
+        registry.add("spring.flyway.user", postgreSQLContainer::getUsername);
+        registry.add("spring.flyway.password", postgreSQLContainer::getPassword);
+    }
 
     @Autowired
     private AppointmentRepository appointmentRepository;
@@ -51,7 +77,7 @@ class AppointmentRepositoryTest extends BaseRepositoryTest {
         List<Appointment> appointmentByMaster = appointmentRepository.getAppointmentByMaster(master);
 
         // THEN
-        assertEquals(4, appointmentByMaster.size());
+        assertEquals(1, appointmentByMaster.size());
     }
 
     @Test
@@ -63,7 +89,7 @@ class AppointmentRepositoryTest extends BaseRepositoryTest {
         List<Appointment> appointmentByClient = appointmentRepository.getAppointmentByClient(client);
 
         // THEN
-        assertEquals(3, appointmentByClient.size());
+        assertEquals(1, appointmentByClient.size());
     }
 
     private static Users getUsers(long id) {
