+++ BeatyPlace/src/test/java/by/beaty/place/repository/UserRepositoryRepositoryTest.java
@@ -6,16 +6,42 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import by.beaty.place.model.Users;
 import by.beaty.place.model.common.Role;
 import by.beaty.place.repository.config.BaseRepositoryTest;
+import java.time.Duration;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.junit.jupiter.Container;
 
-class UserRepositoryTest extends BaseRepositoryTest {
+class UserRepositoryRepositoryTest extends BaseRepositoryTest {
 
     private static final String USERNAME = "client1";
     private static final long USER_ID = 4L;
     private static final String EMAIL = "admin@example.com";
+
+    @Container
+    static PostgreSQLContainer<?> postgreSQLContainer = new PostgreSQLContainer<>("postgres:15")
+            .withDatabaseName("prop")
+            .withUsername("postgres")
+            .withPassword("postgres")
+            .waitingFor(Wait.forListeningPort().withStartupTimeout(Duration.ofSeconds(60)))
+            .withExposedPorts(5432);
+
+    @DynamicPropertySource
+    static void registerPgProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url",
+                () -> String.format("jdbc:postgresql://localhost:%d/prop", postgreSQLContainer.getFirstMappedPort()));
+        registry.add("spring.datasource.username", () -> "postgres");
+        registry.add("spring.datasource.password", () -> "postgres");
+        registry.add("spring.flyway.url", () -> String.format("jdbc:postgresql://localhost:%d/prop",
+                postgreSQLContainer.getFirstMappedPort()));
+        registry.add("spring.flyway.user", postgreSQLContainer::getUsername);
+        registry.add("spring.flyway.password", postgreSQLContainer::getPassword);
+    }
 
     @Autowired
     private UserRepository usersRepository;
