+++ BeatyPlace/src/test/java/by/beaty/place/repository/config/BaseRepositoryTest.java
@@ -1,48 +1,12 @@
 package by.beaty.place.repository.config;
 
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.containers.wait.strategy.Wait;
-import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
+@DataJpaTest
 @Testcontainers
 @Sql(value = "classpath:/sql/initDB.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
 public abstract class BaseRepositoryTest {
 
-    @Container
-    static PostgreSQLContainer<?> postgreSQLContainer = new PostgreSQLContainer<>("postgres:15")
-            .withDatabaseName("prop")
-            .withUsername("postgres")
-            .withPassword("postgres")
-            .waitingFor(Wait.forListeningPort())
-            .withExposedPorts(5432)
-            .withReuse(true);
-
-    @BeforeAll
-    static void startContainer() {
-        postgreSQLContainer.start();
-    }
-
-    @AfterAll
-    static void stopContainer() {
-        postgreSQLContainer.stop();
-    }
-
-    @DynamicPropertySource
-    static void registerPgProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url",
-                () -> String.format("jdbc:postgresql://localhost:%d/prop", postgreSQLContainer.getFirstMappedPort()));
-        registry.add("spring.datasource.username", () -> "postgres");
-        registry.add("spring.datasource.password", () -> "postgres");
-    }
 }
