--- BeatyPlace/src/test/resources/application.yml
@@ -1,12 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:postgresql://localhost:5432/prop
-    username: postgres
-    password: postgres
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    database-platform: org.hibernate.dialect.PostgreSQLDialect
-    hibernate:
-      ddl-auto: update
-  flyway:
-    enabled: true
